# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Uroflow Tracker API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type VoidingEventUrineColor =
  (typeof VoidingEventUrineColor)[keyof typeof VoidingEventUrineColor];

export const VoidingEventUrineColor = {
  pale_yellow: "pale_yellow",
  yellow: "yellow",
  dark_yellow: "dark_yellow",
  orange: "orange",
  dark_orange: "dark_orange",
} as const;

export type VoidingEventAppearanceTagsItem =
  (typeof VoidingEventAppearanceTagsItem)[keyof typeof VoidingEventAppearanceTagsItem];

export const VoidingEventAppearanceTagsItem = {
  clots: "clots",
  flakes: "flakes",
  specks: "specks",
} as const;

export type VoidingEventHematuria =
  (typeof VoidingEventHematuria)[keyof typeof VoidingEventHematuria];

export const VoidingEventHematuria = {
  none: "none",
  visible_hematuria: "visible_hematuria",
  post_drops: "post_drops",
  post_pink: "post_pink",
} as const;

export type VoidingEventUrgency =
  | (typeof VoidingEventUrgency)[keyof typeof VoidingEventUrgency]
  | null;

export const VoidingEventUrgency = {
  none: "none",
  awareness: "awareness",
  urgent: "urgent",
  highly_urgent: "highly_urgent",
  sudden_onset: "sudden_onset",
} as const;

export type VoidingEventPainLocationsItem =
  (typeof VoidingEventPainLocationsItem)[keyof typeof VoidingEventPainLocationsItem];

export const VoidingEventPainLocationsItem = {
  spasm: "spasm",
  perineum: "perineum",
  shaft: "shaft",
  tip: "tip",
} as const;

export type VoidingEventStream =
  | (typeof VoidingEventStream)[keyof typeof VoidingEventStream]
  | null;

export const VoidingEventStream = {
  strong: "strong",
  normal: "normal",
  weak: "weak",
  intermittent: "intermittent",
  dribbling: "dribbling",
} as const;

export interface VoidingEvent {
  id: number;
  voidedAt: string;
  /** @minimum 0 */
  volumeMl: number;
  qmax?: number | null;
  /** @minimum 0 */
  durationSeconds?: number | null;
  urineColor: VoidingEventUrineColor;
  cloudy: boolean;
  appearanceTags?: VoidingEventAppearanceTagsItem[] | null;
  hematuria: VoidingEventHematuria;
  urgency?: VoidingEventUrgency;
  painLocations?: VoidingEventPainLocationsItem[] | null;
  stream?: VoidingEventStream;
  isNocturia: boolean;
  notes?: string | null;
  createdAt: string;
}

export type CreateVoidingEventUrineColor =
  (typeof CreateVoidingEventUrineColor)[keyof typeof CreateVoidingEventUrineColor];

export const CreateVoidingEventUrineColor = {
  pale_yellow: "pale_yellow",
  yellow: "yellow",
  dark_yellow: "dark_yellow",
  orange: "orange",
  dark_orange: "dark_orange",
} as const;

export type CreateVoidingEventAppearanceTagsItem =
  (typeof CreateVoidingEventAppearanceTagsItem)[keyof typeof CreateVoidingEventAppearanceTagsItem];

export const CreateVoidingEventAppearanceTagsItem = {
  clots: "clots",
  flakes: "flakes",
  specks: "specks",
} as const;

export type CreateVoidingEventHematuria =
  (typeof CreateVoidingEventHematuria)[keyof typeof CreateVoidingEventHematuria];

export const CreateVoidingEventHematuria = {
  none: "none",
  visible_hematuria: "visible_hematuria",
  post_drops: "post_drops",
  post_pink: "post_pink",
} as const;

export type CreateVoidingEventUrgency =
  | (typeof CreateVoidingEventUrgency)[keyof typeof CreateVoidingEventUrgency]
  | null;

export const CreateVoidingEventUrgency = {
  none: "none",
  awareness: "awareness",
  urgent: "urgent",
  highly_urgent: "highly_urgent",
  sudden_onset: "sudden_onset",
} as const;

export type CreateVoidingEventPainLocationsItem =
  (typeof CreateVoidingEventPainLocationsItem)[keyof typeof CreateVoidingEventPainLocationsItem];

export const CreateVoidingEventPainLocationsItem = {
  spasm: "spasm",
  perineum: "perineum",
  shaft: "shaft",
  tip: "tip",
} as const;

export type CreateVoidingEventStream =
  | (typeof CreateVoidingEventStream)[keyof typeof CreateVoidingEventStream]
  | null;

export const CreateVoidingEventStream = {
  strong: "strong",
  normal: "normal",
  weak: "weak",
  intermittent: "intermittent",
  dribbling: "dribbling",
} as const;

export interface CreateVoidingEvent {
  voidedAt: string;
  /** @minimum 0 */
  volumeMl: number;
  qmax?: number | null;
  /** @minimum 0 */
  durationSeconds?: number | null;
  urineColor: CreateVoidingEventUrineColor;
  cloudy: boolean;
  appearanceTags?: CreateVoidingEventAppearanceTagsItem[] | null;
  hematuria: CreateVoidingEventHematuria;
  urgency?: CreateVoidingEventUrgency;
  painLocations?: CreateVoidingEventPainLocationsItem[] | null;
  stream?: CreateVoidingEventStream;
  isNocturia: boolean;
  notes?: string | null;
}

export type FluidIntakeEventDrinkTypesItem =
  (typeof FluidIntakeEventDrinkTypesItem)[keyof typeof FluidIntakeEventDrinkTypesItem];

export const FluidIntakeEventDrinkTypesItem = {
  caffeine: "caffeine",
  fizzy: "fizzy",
  acidic: "acidic",
  alcohol: "alcohol",
  neutral: "neutral",
} as const;

export interface FluidIntakeEvent {
  id: number;
  recordedAt: string;
  /** @minimum 0 */
  volumeMl: number;
  drinkTypes?: FluidIntakeEventDrinkTypesItem[] | null;
  createdAt: string;
}

export type CreateFluidIntakeEventDrinkTypesItem =
  (typeof CreateFluidIntakeEventDrinkTypesItem)[keyof typeof CreateFluidIntakeEventDrinkTypesItem];

export const CreateFluidIntakeEventDrinkTypesItem = {
  caffeine: "caffeine",
  fizzy: "fizzy",
  acidic: "acidic",
  alcohol: "alcohol",
  neutral: "neutral",
} as const;

export interface CreateFluidIntakeEvent {
  recordedAt: string;
  /** @minimum 0 */
  volumeMl: number;
  drinkTypes?: CreateFluidIntakeEventDrinkTypesItem[] | null;
}

export interface StatsPeriod {
  count: number;
  avgVolumeMl: number | null;
  totalVolumeMl: number;
  bloodIncidents: number;
  avgDurationSeconds: number | null;
}

export interface VoidingStats {
  last7Days: StatsPeriod;
  last30Days: StatsPeriod;
  totalRecords: number;
}

export interface FluidBalancePoint {
  date: string;
  cumulativeIntakeMl: number;
  cumulativeVoidedMl: number;
  netBalanceMl: number;
}

export interface NocturiaNight {
  date: string;
  count: number;
}

export interface NocturiaStats {
  totalEvents: number;
  avgPerNight: number;
  nights: NocturiaNight[];
}

export interface UrgencyVsVolume {
  urgency: string;
  avgVolumeMl: number;
  count: number;
}

export interface ReportData {
  fluidBalance: FluidBalancePoint[];
  nocturia: NocturiaStats;
  urgencyVsVolume: UrgencyVsVolume[];
}