# lib/api-zod/src/generated/types/createFluidIntakeEvent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Uroflow Tracker API
 * OpenAPI spec version: 0.2.0
 */
import type { CreateFluidIntakeEventDrinkTypesItem } from "./createFluidIntakeEventDrinkTypesItem";

export interface CreateFluidIntakeEvent {
  recordedAt: Date;
  /** @minimum 0 */
  volumeMl: number;
  drinkTypes?: CreateFluidIntakeEventDrinkTypesItem[] | null;
}