# lib/api-zod/src/generated/types/fluidIntakeEvent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Uroflow Tracker API
 * OpenAPI spec version: 0.2.0
 */
import type { FluidIntakeEventDrinkTypesItem } from "./fluidIntakeEventDrinkTypesItem";

export interface FluidIntakeEvent {
  id: number;
  recordedAt: Date;
  /** @minimum 0 */
  volumeMl: number;
  drinkTypes?: FluidIntakeEventDrinkTypesItem[] | null;
  createdAt: Date;
}