# lib/api-zod/src/generated/types/nocturiaStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Uroflow Tracker API
 * OpenAPI spec version: 0.2.0
 */
import type { NocturiaNight } from "./nocturiaNight";

export interface NocturiaStats {
  totalEvents: number;
  avgPerNight: number;
  nights: NocturiaNight[];
}