# lib/api-zod/src/generated/types/reportData.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Uroflow Tracker API
 * OpenAPI spec version: 0.2.0
 */
import type { FluidBalancePoint } from "./fluidBalancePoint";
import type { NocturiaStats } from "./nocturiaStats";
import type { UrgencyVsVolume } from "./urgencyVsVolume";

export interface ReportData {
  fluidBalance: FluidBalancePoint[];
  nocturia: NocturiaStats;
  urgencyVsVolume: UrgencyVsVolume[];
}