# lib/api-zod/src/generated/types/statsPeriod.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Uroflow Tracker API
 * OpenAPI spec version: 0.2.0
 */

export interface StatsPeriod {
  count: number;
  avgVolumeMl: number | null;
  totalVolumeMl: number;
  bloodIncidents: number;
  avgDurationSeconds: number | null;
}