# lib/api-zod/src/generated/types/urgencyVsVolume.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Uroflow Tracker API
 * OpenAPI spec version: 0.2.0
 */

export interface UrgencyVsVolume {
  urgency: string;
  avgVolumeMl: number;
  count: number;
}