# lib/api-zod/src/generated/types/voidingEvent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Uroflow Tracker API
 * OpenAPI spec version: 0.2.0
 */
import type { VoidingEventAppearanceTagsItem } from "./voidingEventAppearanceTagsItem";
import type { VoidingEventHematuria } from "./voidingEventHematuria";
import type { VoidingEventPainLocationsItem } from "./voidingEventPainLocationsItem";
import type { VoidingEventStream } from "./voidingEventStream";
import type { VoidingEventUrgency } from "./voidingEventUrgency";
import type { VoidingEventUrineColor } from "./voidingEventUrineColor";

export interface VoidingEvent {
  id: number;
  voidedAt: Date;
  /** @minimum 0 */
  volumeMl: number;
  qmax?: number | null;
  /** @minimum 0 */
  durationSeconds?: number | null;
  urineColor: VoidingEventUrineColor;
  cloudy: boolean;
  appearanceTags?: VoidingEventAppearanceTagsItem[] | null;
  hematuria: VoidingEventHematuria;
  urgency?: VoidingEventUrgency;
  painLocations?: VoidingEventPainLocationsItem[] | null;
  stream?: VoidingEventStream;
  isNocturia: boolean;
  notes?: string | null;
  createdAt: Date;
}